+++ include/minizinc/flatten_internal.hh
@@ -294,7 +294,8 @@ class EnvI {
   friend class TypeInst;
   template <bool ignoreVarDecl>
   friend class Typer;
-  friend KeepAlive add_coercion(EnvI& env, Model* m, Expression* e, const Type& funarg_t);
+  friend KeepAlive add_coercion(EnvI& env, Model* m, Expression* e, const Location& loc_default,
+                                const Type& funarg_t);
   friend Type type_from_tmap(EnvI& env, TypeInst* ti,
                              const ASTStringMap<std::pair<Type, bool>>& tmap);